# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News API",
        "description": "API для керування новинами",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8082/api",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/profile": {
            "get": {
                "summary": "Отримати профіль",
                "operationId": "c6a79007a56511faa6f48397a2d15867",
                "responses": {
                    "200": {
                        "description": "Дані користувача"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "summary": "Оновити профіль",
                "operationId": "74599296b1816be41ad4228975229966",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "New Name"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "new@email.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Профіль оновлено"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Видалити акаунт",
                "operationId": "146e3d0726049246ab0e167e0e7ad1bb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Акаунт видалено"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profile/avatar": {
            "post": {
                "summary": "Оновити аватар",
                "operationId": "fd54c82c040d44a02c94a38b5eafa899",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "avatar"
                                ],
                                "properties": {
                                    "avatar": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Аватар оновлено"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/history": {
            "get": {
                "summary": "Історія переглядів",
                "operationId": "bbef1e618e15f097bf74c48446f72242",
                "responses": {
                    "200": {
                        "description": "Список переглядів"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/history/clear": {
            "delete": {
                "summary": "Очистити історію",
                "operationId": "981b9847de6d065b2dd396a0786e5f85",
                "responses": {
                    "200": {
                        "description": "Історію очищено"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}